+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 * <kbd>C-c M-r</kbd> will rotate and display the current nREPL connection.
 * Setting the variable `nrepl-buffer-name-show-port` will display the port on which the nRepl server is running.
 * nRepl buffer name uses project directory name; `*nrepl*` will appear as `*nrepl project-directory-name*`.
-* <kbd>C-c C-z</kbd> will select the clojure buffer based on the current namespace.
-* <kbd>C-u C-u C-c C-z</kbd> will select the clojure buffer based on a user directory prompt.
+* <kbd>C-c C-Z</kbd> will select the nrepl buffer based on the current namespace.
+* <kbd>C-u C-c C-Z</kbd> will select the nrepl buffer based on a user project directory prompt.
 
 ### Bugs fixed